refactor: 서비스 레이어 도메인별 구조화
