+++ backend/demo/src/main/java/com/sesac/solbid/controller/PaymentController.java
@@ -2,7 +2,7 @@ package com.sesac.solbid.controller;
 
 import com.sesac.solbid.dto.payment.request.PaymentPrepareRequest;
 import com.sesac.solbid.dto.payment.response.PaymentPrepareResponse;
-import com.sesac.solbid.service.PaymentService;
+import com.sesac.solbid.service.payment.PaymentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
