+++ backend/demo/src/main/java/com/sesac/solbid/controller/PointController.java
@@ -2,12 +2,11 @@ package com.sesac.solbid.controller;
 
 import com.sesac.solbid.domain.User;
 import com.sesac.solbid.dto.payment.response.PointSummaryResponse;
-import com.sesac.solbid.service.PointService;
+import com.sesac.solbid.service.point.PointService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
