+++ backend/demo/src/main/java/com/sesac/solbid/service/payment/PaymentRecordQueryService.java
@@ -1,4 +1,4 @@
-package com.sesac.solbid.service;
+package com.sesac.solbid.service.payment;
 
 import com.sesac.solbid.dto.payment.request.PaymentRecordSearchRequest;
 import com.sesac.solbid.dto.payment.response.PageResponse;
