+++ backend/demo/src/main/java/com/sesac/solbid/service/payment/PaymentServiceImpl.java
@@ -1,4 +1,4 @@
-package com.sesac.solbid.service;
+package com.sesac.solbid.service.payment;
 
 import com.sesac.solbid.domain.Payments;
 import com.sesac.solbid.domain.User;
