+++ backend/demo/src/main/java/com/sesac/solbid/service/point/PointService.java
@@ -1,4 +1,4 @@
-package com.sesac.solbid.service;
+package com.sesac.solbid.service.point;
 
 import com.sesac.solbid.dto.payment.response.PointSummaryResponse;
