+++ backend/demo/src/main/java/com/sesac/solbid/service/s3/S3Service.java
@@ -1,4 +1,4 @@
-package com.sesac.solbid.service;
+package com.sesac.solbid.service.s3;
 
 public interface S3Service {
