+++ backend/demo/src/main/java/com/sesac/solbid/service/s3/S3StorageServiceImpl.java
@@ -1,4 +1,4 @@
-package com.sesac.solbid.service;
+package com.sesac.solbid.service.s3;
 
 import com.sesac.solbid.exception.CustomException;
 import com.sesac.solbid.exception.ErrorCode;
